update user id foreignkey

+++ database/migrations/2021_10_13_085659_create_users_courses_table.php
@@ -19,7 +19,7 @@ class CreateUsersCoursesTable extends Migration
             $table->boolean('checked')->default(0);
             $table->boolean('is_true')->default(0);
             $table->char('course_id', 10);
-            $table->bigIncrements('user_id');
+            $table->integer('user_id');
             $table->foreign('course_id')->references('id')->on('courses')->onDelete('RESTRICT')->onUpdate('CASCADE');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('RESTRICT')->onUpdate('CASCADE');
             $table->timestamps();
